+++ src_new/triggers/levi_ataxia/for_levi/leviticus/mnemosyne/001_Run_Start.lua
@@ -68,6 +68,7 @@ psionRazorClarity = false  -- boons reset each run
 psionMindbreak = false  -- boons reset each run
 psionPsiwave = false  -- boons reset each run
 psionEarthquake = false  -- boons reset each run
+psionProphet = false  -- boons reset each run
 dragonMightSycaerunax = false  -- boons reset each run
 dragonRampage = false  -- boons reset each run
 dwFlashforward = false  -- boons reset each run